fix: push data to server
Previously newly entered data was deleted for failed sync. now it has
been fixed.

Change the sync message system. Now the synchronizing message will close
after the end of the sync. There is new animation on the message

The count of not synced data will show on top of the sync button.