fix: sync issue (#79)
* fix: the sync issue getting data from server

now it will try to 5 times recall the failed request.

previously the data was not got if some request was failed. now the application will retry the request up to 5
times if some request is failed.

* fix: push data to server

Previously newly entered data was deleted for failed sync. now it has
been fixed.

Change the sync message system. Now the synchronizing message will close
after the end of the sync. There is new animation on the message

The count of not synced data will show on top of the sync button.

* fix: createTheme error on the startup the application

* fix: sync message issue